已经能编译成功

+++ includes/smooth.h
@@ -65,15 +65,15 @@ namespace HybridAStar
         //碰撞图
         DynamicVoronoi voronoi;
         //惩罚边界
-        double kappaMax = 1. / (param::rmin * 1.1);     //最大曲率
-        double obsDMax = param::obsPenaMax;              //障碍物惩罚的最大距离
-        double voronoiMax = param::obsPenaMax;          //影响voronoi场的最大距离
+        float kappaMax = 1. / (param::rmin * 1.1);     //最大曲率
+        float obsDMax = param::obsPenaMax;              //障碍物惩罚的最大距离
+        float voronoiMax = param::obsPenaMax;          //影响voronoi场的最大距离
         //权重系数
-        double alpha = 0.1;                          //梯度更新衰减率
-        double wObstacle = 0.1;                         //障碍物项权重
-        double wCurvature = 0.1;                        //曲率项权重
-        double wSmoothness = 0.5;                       //顺滑项权重
-        double wVoronoi = 0;                            //voronoi项权重
+        float alpha = 0.1;                          //梯度更新衰减率
+        float wObstacle = 0.1;                         //障碍物项权重
+        float wCurvature = 0.1;                        //曲率项权重
+        float wSmoothness = 0.5;                       //顺滑项权重
+        float wVoronoi = 0;                            //voronoi项权重
         //地图尺寸
         int width;
         int height;
